# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Arduino IDE API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Project {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  code: string;
  board: string;
  createdAt: string;
  updatedAt: string;
}

export interface ProjectInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  code?: string;
  board?: string;
}

export interface ProjectUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  code?: string;
  board?: string;
}

export interface CompileInput {
  code: string;
  board: string;
}

export interface CompileResult {
  success: boolean;
  output: string;
  errors: string[];
  warnings: string[];
  /** @nullable */
  binarySize?: number | null;
}

export interface GenerateCodeInput {
  prompt: string;
  /** @nullable */
  context?: string | null;
  board: string;
}

export interface FixCodeInput {
  code: string;
  error: string;
  board: string;
}

export interface ExplainCodeInput {
  code: string;
}

export interface ApiError {
  error: string;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface OpenaiConversationInput {
  title: string;
}

export interface OpenaiTextMessageInput {
  content: string;
}

export interface OpenaiVoiceInput {
  /** Base64-encoded audio data */
  audio: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export interface OpenaiError {
  error: string;
}