# lib/api-zod/src/generated/types/project.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Arduino IDE API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Project {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  code: string;
  board: string;
  createdAt: Date;
  updatedAt: Date;
}