# lib/api-zod/src/generated/types/projectUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Arduino IDE API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  code?: string;
  board?: string;
}